Remove YAML setup from Withings
